+++ ts-definitions/tests.d.ts
@@ -269,16 +269,20 @@ declare namespace BrowserTests {
 		setup(page: Playwright.Page, context: Playwright.BrowserContext, options?: SetupOptions): Promise<void>;
 
 		/**
-		 * Collects all of the console.log calls on the page
+		 * Intercepts and collects all invoking of `console` methods on the specified page.
+		 * Mind, the intercepted callings aren't be shown a console till you invoke the `printPageLogs` method.
 		 *
 		 * @param page
 		 */
 		collectPageLogs(page: Playwright.Page);
 
 		/**
-		 * Prints all of the collected console.log calls on the page into the terminal
+		 * Prints all of the intercepted console invokes to a console
+		 *
+		 * @param page
+		 *
 		 */
-		printPageLogs();
+		printPageLogs(page: Playwright.Page);
 
 		/**
 		 * Reloads the page and waits until `requestIdleCallback`