+++ CHANGELOG.md
@@ -17,6 +17,10 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * Added `TestParams` interface for `BrowserTests` `test.d.ts`
 
+#### :rocket: New Feature
+
+* [https://github.com/V4Fire/Client/issues/168](Added `optionsResolver` into `core/dom/image`)
+
 ## v3.8.3 (2021-10-26)
 
 #### :bug: Bug Fix