+++ CHANGELOG.md
@@ -19,7 +19,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :rocket: New Feature
 
-* [https://github.com/V4Fire/Client/issues/168](Added `optionsResolver` into `core/dom/image`)
+* [Added `optionsResolver` into `core/dom/image`](https://github.com/V4Fire/Client/issues/168)
 
 ## v3.8.3 (2021-10-26)